created start of Capture the Flag game with new state interface
