+++ src/main/java/org/braekpo1nt/mctmanager/games/game/capturetheflag/states/CaptureTheFlagState.java
@@ -1,0 +1,25 @@
+package org.braekpo1nt.mctmanager.games.game.capturetheflag.states;
+
+import org.bukkit.entity.Player;
+import org.bukkit.event.entity.EntityDamageEvent;
+import org.bukkit.event.entity.FoodLevelChangeEvent;
+import org.bukkit.event.inventory.InventoryClickEvent;
+import org.bukkit.event.player.PlayerMoveEvent;
+
+public interface CaptureTheFlagState {
+    void onParticipantJoin(Player participant);
+    void onParticipantQuit(Player participant);
+    void initializeParticipant(Player participant);
+    void resetParticipant(Player participant);
+    
+    default void cancelAllTasks() {
+        // do nothing
+    }
+    
+    // event handlers
+    void onPlayerDamage(EntityDamageEvent event);
+    void onPlayerLoseHunger(FoodLevelChangeEvent event);
+    void onPlayerMove(PlayerMoveEvent event);
+    void onClickInventory(InventoryClickEvent event);
+    
+}
